Implement SelectQueries for Postgres and move model statistics to ModelInferenceQueries
Add model provider statistics support for Postgres backend:

- Create materialized view migration for model_provider_statistics
- Implement SelectQueries for PostgresConnectionInfo:
  - count_distinct_models_used
  - get_model_usage_timeseries
  - get_model_latency_quantiles (22 quantiles via percentile_cont)
  - query_episode_table
  - query_episode_table_bounds
- Add sqlx::FromRow derives to EpisodeByIdRow and TableBoundsWithCount
- Implement SelectQueries for DelegatingDatabaseConnection with feature flag routing
- Update model and episode endpoint handlers to use DelegatingDatabaseConnection

Move model statistics methods to ModelInferenceQueries:

- Move count_distinct_models_used, get_model_usage_timeseries, get_model_latency_quantiles from SelectQueries to ModelInferenceQueries trait
- Update ClickHouse and Postgres implementations
- Update DelegatingDatabaseConnection and endpoint handlers
- SelectQueries now only contains episode-related methods
