+++ tensorzero-core/src/db/clickhouse/mock_clickhouse_connection_info.rs
@@ -17,7 +17,9 @@ use crate::db::inferences::{
 };
 use crate::db::model_inferences::{MockModelInferenceQueries, ModelInferenceQueries};
 use crate::db::stored_datapoint::StoredDatapoint;
-use crate::db::{ConfigQueries, MockConfigQueries};
+use crate::db::{
+    ConfigQueries, MockConfigQueries, ModelLatencyDatapoint, ModelUsageTimePoint, TimeWindow,
+};
 use crate::error::Error;
 use crate::inference::types::StoredModelInference;
 use crate::inference::types::{ChatInferenceDatabaseInsert, JsonInferenceDatabaseInsert};
@@ -252,5 +254,35 @@ impl ModelInferenceQueries for MockClickHouseConnectionInfo {
         self.model_inference_queries
             .insert_model_inferences(rows)
             .await
+    }
+
+    async fn count_distinct_models_used(&self) -> Result<u32, Error> {
+        self.model_inference_queries
+            .count_distinct_models_used()
+            .await
+    }
+
+    async fn get_model_usage_timeseries(
+        &self,
+        time_window: TimeWindow,
+        max_periods: u32,
+    ) -> Result<Vec<ModelUsageTimePoint>, Error> {
+        self.model_inference_queries
+            .get_model_usage_timeseries(time_window, max_periods)
+            .await
+    }
+
+    async fn get_model_latency_quantiles(
+        &self,
+        time_window: TimeWindow,
+    ) -> Result<Vec<ModelLatencyDatapoint>, Error> {
+        self.model_inference_queries
+            .get_model_latency_quantiles(time_window)
+            .await
+    }
+
+    fn get_model_latency_quantile_function_inputs(&self) -> &[f64] {
+        // Return ClickHouse quantiles since this is a mock for ClickHouse
+        super::migration_manager::migrations::migration_0037::QUANTILES
     }
 }
