+++ tensorzero-core/src/db/clickhouse/model_inferences.rs
@@ -6,8 +6,10 @@ use async_trait::async_trait;
 use uuid::Uuid;
 
 use super::ClickHouseConnectionInfo;
+use super::migration_manager::migrations::migration_0037::{QUANTILES, quantiles_sql_args};
 use super::table_name::TableName;
 use crate::db::model_inferences::ModelInferenceQueries;
+use crate::db::{ModelLatencyDatapoint, ModelUsageTimePoint, TimeWindow};
 use crate::error::{Error, ErrorDetails};
 use crate::inference::types::StoredModelInference;
 
@@ -88,6 +90,166 @@ impl ModelInferenceQueries for ClickHouseConnectionInfo {
 
         self.write_batched(&serialized, TableName::ModelInference)
             .await
+    }
+
+    async fn count_distinct_models_used(&self) -> Result<u32, Error> {
+        let query =
+            "SELECT toUInt32(uniqExact(model_name)) FROM ModelProviderStatistics".to_string();
+        let response = self.run_query_synchronous_no_params(query).await?;
+        response
+            .response
+            .trim()
+            .lines()
+            .next()
+            .ok_or_else(|| {
+                Error::new(ErrorDetails::ClickHouseDeserialization {
+                    message: "No result".to_string(),
+                })
+            })?
+            .parse()
+            .map_err(|e: std::num::ParseIntError| {
+                Error::new(ErrorDetails::ClickHouseDeserialization {
+                    message: e.to_string(),
+                })
+            })
+    }
+
+    /// Retrieves a timeseries of model usage data.
+    /// This will return max_periods complete time periods worth of data if present
+    /// as well as the current time period's data.
+    /// So there are at most max_periods + 1 time periods worth of data returned.
+    async fn get_model_usage_timeseries(
+        &self,
+        time_window: TimeWindow,
+        max_periods: u32,
+    ) -> Result<Vec<ModelUsageTimePoint>, Error> {
+        // TODO: probably factor this out into common code as other queries will likely need similar logic
+        // NOTE: this filter pattern will likely include some extra data since the current period is likely incomplete.
+        let (time_grouping, time_filter) = match time_window {
+            TimeWindow::Minute => (
+                "toStartOfMinute(minute)",
+                format!(
+                    "minute >= (SELECT max(toStartOfMinute(minute)) FROM ModelProviderStatistics) - INTERVAL {max_periods} MINUTE"
+                ),
+            ),
+            TimeWindow::Hour => (
+                "toStartOfHour(minute)",
+                format!(
+                    "minute >= (SELECT max(toStartOfHour(minute)) FROM ModelProviderStatistics) - INTERVAL {max_periods} HOUR"
+                ),
+            ),
+            TimeWindow::Day => (
+                "toStartOfDay(minute)",
+                format!(
+                    "minute >= (SELECT max(toStartOfDay(minute)) FROM ModelProviderStatistics) - INTERVAL {max_periods} DAY"
+                ),
+            ),
+            TimeWindow::Week => (
+                "toStartOfWeek(minute)",
+                format!(
+                    "minute >= (SELECT max(toStartOfWeek(minute)) FROM ModelProviderStatistics) - INTERVAL {max_periods} WEEK"
+                ),
+            ),
+            TimeWindow::Month => (
+                "toStartOfMonth(minute)",
+                format!(
+                    "minute >= (SELECT max(toStartOfMonth(minute)) FROM ModelProviderStatistics) - INTERVAL {max_periods} MONTH"
+                ),
+            ),
+            TimeWindow::Cumulative => (
+                "toDateTime('1970-01-01 00:00:00')",
+                "1 = 1".to_string(), // No time filter for cumulative
+            ),
+        };
+
+        let query = format!(
+            r"
+            SELECT
+                formatDateTime({time_grouping}, '%Y-%m-%dT%H:%i:%SZ') as period_start,
+                model_name,
+                sumMerge(total_input_tokens) as input_tokens,
+                sumMerge(total_output_tokens) as output_tokens,
+                countMerge(count) as count
+            FROM ModelProviderStatistics
+            WHERE {time_filter}
+            GROUP BY period_start, model_name
+            ORDER BY period_start DESC, model_name
+            FORMAT JSONEachRow
+            ",
+        );
+
+        let response = self.run_query_synchronous_no_params(query).await?;
+
+        // Deserialize the results into ModelUsageTimePoint
+        response
+            .response
+            .trim()
+            .lines()
+            .map(|row| {
+                serde_json::from_str(row).map_err(|e| {
+                    Error::new(ErrorDetails::ClickHouseDeserialization {
+                        message: e.to_string(),
+                    })
+                })
+            })
+            .collect::<Result<Vec<_>, _>>()
+    }
+
+    async fn get_model_latency_quantiles(
+        &self,
+        time_window: TimeWindow,
+    ) -> Result<Vec<ModelLatencyDatapoint>, Error> {
+        let time_filter = match time_window {
+            TimeWindow::Minute => {
+                "minute >= (SELECT max(minute) FROM ModelProviderStatistics) - INTERVAL 1 MINUTE"
+            }
+            TimeWindow::Hour => {
+                "minute >= (SELECT max(minute) FROM ModelProviderStatistics) - INTERVAL 1 HOUR"
+            }
+            TimeWindow::Day => {
+                "minute >= (SELECT max(minute) FROM ModelProviderStatistics) - INTERVAL 1 DAY"
+            }
+            TimeWindow::Week => {
+                "minute >= (SELECT max(minute) FROM ModelProviderStatistics) - INTERVAL 1 WEEK"
+            }
+            TimeWindow::Month => {
+                "minute >= (SELECT max(minute) FROM ModelProviderStatistics) - INTERVAL 1 MONTH"
+            }
+            TimeWindow::Cumulative => "1 = 1",
+        };
+        let qs = quantiles_sql_args();
+        let query = format!(
+            r"
+            SELECT
+                model_name,
+                quantilesTDigestMerge({qs})(response_time_ms_quantiles) AS response_time_ms_quantiles,
+                quantilesTDigestMerge({qs})(ttft_ms_quantiles) AS ttft_ms_quantiles,
+                countMerge(count) as count
+            FROM ModelProviderStatistics
+            WHERE {time_filter}
+            GROUP BY model_name
+            ORDER BY model_name
+            FORMAT JSONEachRow
+            ",
+        );
+        let response = self.run_query_synchronous_no_params(query).await?;
+        // Deserialize the results into ModelLatencyDatapoint
+        response
+            .response
+            .trim()
+            .lines()
+            .map(|row| {
+                serde_json::from_str(row).map_err(|e| {
+                    Error::new(ErrorDetails::ClickHouseDeserialization {
+                        message: e.to_string(),
+                    })
+                })
+            })
+            .collect::<Result<Vec<_>, _>>()
+    }
+
+    fn get_model_latency_quantile_function_inputs(&self) -> &[f64] {
+        QUANTILES
     }
 }
